% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{execute_sql_file}
\alias{execute_sql_file}
\title{execute sql file}
\usage{
execute_sql_file(push_connection, sql_file_path)
}
\arguments{
\item{push_connection}{the connection}

\item{sql_file_path}{the list}
}
\value{
nothing
}
\description{
executes sql file where each block of code to be executed ends with ";"
}
\examples{
execute_sql_file(my_conn, my_file)
}
