% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{execute_sql_list}
\alias{execute_sql_list}
\title{execute list of sql statements}
\usage{
execute_sql_list(push_connection, list_of_sql_statements)
}
\arguments{
\item{push_connection}{the connection}

\item{list_of_sql_statements}{the list}
}
\value{
nothing
}
\description{
executes list of sql statements
}
\examples{
execute_sql_list(my_conn, my_list)
}
