% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{extract}
\alias{extract}
\title{extract function for read ini file function}
\usage{
extract(regexp, x)
}
\arguments{
\item{regexp}{blah}

\item{x}{blah}
}
\value{
blah
}
\description{
used by read ini file function and not intended for use as a stand-alone fuinction
}
\examples{
section = extract(header, l)
}
