% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{pull_data_from_sql_server}
\alias{pull_data_from_sql_server}
\title{pull data from sql server}
\usage{
pull_data_from_sql_server(pull_connection, sql_statement)
}
\arguments{
\item{pull_connection}{pull connection}

\item{sql_statement}{sql statement}
}
\value{
df dataframe
}
\description{
pulls data from sql server
}
\examples{
df <- pull_data_from_sql_server(pull_conn, sql_statement)
}
