% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{push_data_to_sql_server}
\alias{push_data_to_sql_server}
\title{push data to sql server}
\usage{
push_data_to_sql_server(
  push_connection,
  table_name,
  dataframe,
  append_or_overwrite
)
}
\arguments{
\item{push_connection, }{the connection}

\item{table_name}{the table}

\item{dataframe}{the dataframe}

\item{append_or_overwrite}{either "append" or "overwrite"}
}
\value{
nothing
}
\description{
pushes data to sql server
}
\examples{
push_data_to_sql_server(my_conn, table_name, dataframe, append_or_overwrite)
}
