% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{push_sql_statement}
\alias{push_sql_statement}
\title{push sql to sql server}
\usage{
push_sql_statement(push_connection, sql_statement)
}
\arguments{
\item{push_connection}{the connection}

\item{sql_statement}{the sql}
}
\value{
nothing
}
\description{
pushes sql statement to sql server
}
\examples{
push_sql_statement(my_conn, my_sql)
}
