% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_first_line_of_text_file}
\alias{read_first_line_of_text_file}
\title{read first line of text file}
\usage{
read_first_line_of_text_file(text_file_path)
}
\arguments{
\item{text_file_path}{text file path}
}
\value{
first_line first line
}
\description{
reads first line of text file
}
\examples{
first_line <- read_first_line_of_text_file(text_file_path)
}
