% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_ini_file}
\alias{read_ini_file}
\title{read ini file}
\usage{
read_ini_file(ini_file_path)
}
\arguments{
\item{ini_file_path}{ini file path}
}
\value{
list of lists
}
\description{
reads ini file as list of lists
}
\examples{
ini_file <- read_ini_file(ini_file_path)
}
