% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_text_file_as_list}
\alias{read_text_file_as_list}
\title{read text file as list}
\usage{
read_text_file_as_list(text_file_path, list_separator)
}
\arguments{
\item{text_file_path}{text file path}

\item{list_separator}{separator e.g. ";" for sql statements / code blocks}
}
\value{
my_list list
}
\description{
reads text file as list with a given separator
}
\examples{
my_list <- read_text_file_as_list(text_file_path, list_separator)
}
