% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{sql_server_push_connection}
\alias{sql_server_push_connection}
\title{connect to sql server to push data}
\usage{
sql_server_push_connection(data_source_name, config_file_path)
}
\arguments{
\item{data_source_name}{the name of the data source in the ini file}

\item{config_file_path}{the path to the ini file}
}
\value{
db_conn the connection to sql server
}
\description{
uses an ini file to connect to sql server
}
\examples{
push_conn <- sql_server_push_connection(data_source_name, config_file_path)
}
